Does not build
I re-created all the DMX fixtures so they were all much neater and then attempted to create a new class for an 8 colour mix. This caused the SainsburyTheatreVis Module to fail to build. I restarted the engine and now cannot get the file to load. Shit. Commiting now so that if I make more mistakes I can go back to here.