FOH Colours
Ended up changing the colours of the FOH Lustrs in EOS rather than Unreal. Added some files in this commit around the research of how you could potentially take a 7 colour array and convert it to RGB, but I don't understand how Unreal is doing it's colour math well enough. In theory the amber chips should work in fixtures, reverse engineer that and you'll work out what needs to be done.